feat: add support for packageManager field
Node 16.9.0 added support for a packageManager field in package.json.
This field specifies which package manager is expected to be used
in the project. If the field is present, we can use that instead
of checking the lockfiles.

Note: `bun` is not a valid package manager in the validation regex of the
field.